Rename eigenvalues module to point module
